@@ -1,8 +1,6 @@
 package kg.tech.order.entities;
 
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 import lombok.experimental.FieldDefaults;
 import org.hibernate.annotations.Type;
 
@@ -11,6 +9,8 @@ import javax.persistence.*;
 @Entity
 @Getter
 @Setter
+@AllArgsConstructor
+@NoArgsConstructor
 @Table(name = "images")
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class Image {
